outsource ClockFlags
